am adaugat "mecanica" de use a itemelor, itemele cumparate din inventar se pot folosi inainte de lupta, poate nu e destul de clar cu ce se cumpara ce item, evident nu am destule iteme de fiecare tip in jsonuri (o sa am) acum chiar playerii iau damage - totusi as adauga o mecanica de rest sa refaca hp la un anumit timp ca tot retin last_login si de asemenea inca o modalitate ceva (poate un typer) care sa ajute sa faca playerul rost de currency si in alt mod inafara de cel in care nu e pe joc :) ... conversia intre currency uri si ea e aiurea e prea logaritmica momentan -3-

+++ headers/cheating_sheet.cpp
@@ -51,7 +51,7 @@ void cheating_sheet::use(player &p) {
     std::random_device rd;
     std::mt19937 gen(rd());
     std::uniform_real_distribution<> dis(0.0, 1.0);
-    float roll = dis(gen);
+    double roll = dis(gen);
 
     std::cout << "(Success Chance: " << this->success_chance * 100 << "%)\n";
